pyproject.toml
[project]
name = "koreo"
version = "0.3.2"
description = "Tooling package for Koreo"
authors = [
    {name = "Robert Kluin", email = "robert.kluin@realkinetic.com"},
    {name = "Eric Larssen", email = "eric.larssen@realkinetic.com"},
    {name = "Tyler Treat", email = "tyler.treat@realkinetic.com"},
]
dependencies = [
    "cel-python==0.2.0",
    "pygls==2.0.0a2",
    "pyyaml==6.0.2",
    "ruamel.yaml>=0.17.0",
    "koreo-core==0.1.11",
    "colorist>=1.8.3",
    "kr8s>=0.20.6",
]

requires-python = ">=3.13"
readme = "README.md"
license = {text = "Apache-2.0"}
    
[project.urls]
Homepage = "https://koreo.dev"

[project.scripts]
koreo-ls = "server:main"
koreo = "cli.__main__:main"

[tool.pdm.scripts]
lint = "ruff check src/"
lint-fix = "ruff check src/ --fix"
format = "ruff format src/"
test = "pytest"
test-cov = "pytest --cov=src --cov-report=term-missing"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
package-dir = "src"
includes = ["src/koreo_tooling/**/*", "src/cli", "src/server.py"]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"
addopts = [
  "-v",
  "--import-mode=importlib",
  "--cov=src",
  "--cov-branch",
  "--cov-report=term-missing",
]

[dependency-groups]
dev = [
    "pyright==1.1.391",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "colorist>=1.8.3",
    "ruff>=0.8.0",
]

[tool.ruff]
# Python version target
target-version = "py313"

# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default
# Also enable isort (`I`), pyupgrade (`UP`), and flake8-bugbear (`B`)
lint.select = ["E", "F", "I", "UP", "B"]

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".pytest_cache",
]

# Set line length to 80 characters
line-length = 80
indent-width = 4

[tool.ruff.lint.isort]
# Configure import sorting
known-first-party = ["koreo_tooling", "cli", "server"]

[tool.ruff.lint.per-file-ignores]
"src/server.py" = ["E402"]  # Allow imports not at top of file for server initialization

[tool.ruff.format]
# Formatting configuration
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
